+++ src/components/popups/external-transfer-popup/steps/AddInfo.vue
@@ -1,30 +1,28 @@
 <template>
   <div>
     <p class="popup-text">{{ $t("externalTransfer.MoreInfo") }}</p>
-    <Textarea
-      @input="onInfoChange"
-      :placeholder="$t('externalTransfer.Describe')"
-    />
+    <Textarea :reset="resetInfo" @input="onInfoChange" :placeholder="$t('externalTransfer.Describe')" />
     <div class="space-between">
-      <div class="select-container">
+      <div :class="[classification && classification.length < 10 ? 'select-classfier' : 'select-classfier-max']">
         <Select
+          :reset="resetSelect"
           :items="classifications"
           background="transparent"
           @change="onClassificationChange"
           :placeholder="$t('externalTransfer.ChooseClassification')"
         />
       </div>
-      <p v-if="classification === 'סודי ביותר'" class="top-secret">
-        {{ $t("externalTransfer.TopSecret") }}
+      <p v-if="classification && !isClassificationEnable" class="disable-classification">
+        {{
+          $t("externalTransfer.DisableClassification", {
+            classificationName: classification,
+          })
+        }}
       </p>
     </div>
 
     <v-card-actions class="popup-confirm">
-      <SubmitButton
-        @click="onConfirm"
-        :label="$t('buttons.Continue')"
-        :disabled="disabled"
-      />
+      <SubmitButton @click="onConfirm" :label="$t('buttons.Continue')" :disabled="disabled" />
       <TextButton @click="$emit('back')" :label="$t('buttons.Back')" />
     </v-card-actions>
   </div>
@@ -44,8 +42,34 @@ export default {
       info: "",
       classification: undefined,
       disabled: true,
-      classifications: this.$t("externalTransfer.Classifications"),
+      classifications: [],
+      isClassificationEnable: true,
+      resetSelect: null,
+      resetInfo: null,
     };
+  },
+  props: {
+    reset: Boolean,
+    networkDest: String,
+  },
+  watch: {
+    reset() {
+      this.resetSelect = null;
+      this.resetInfo = null;
+      this.disabled = true;
+      this.classification = null;
+      this.isClassificationEnable = true;
+      this.classifications = [];
+    },
+    networkDest() {
+      this.classifications =
+        this.networkDest == undefined ? [] : this.$t(`externalTransfer.Classifications.${this.networkDest}`);
+      this.classification = null;
+      this.info = "";
+      this.resetSelect = null;
+      this.resetInfo = null;
+      this.disabled = true;
+    },
   },
   methods: {
     onConfirm() {
@@ -53,23 +77,37 @@ export default {
     },
     onInfoChange(value) {
       this.info = value;
+      this.resetInfo = value;
       this.toggleDisabled();
     },
     onClassificationChange(value) {
       this.classification = value;
+      this.resetSelect = value;
       this.toggleDisabled();
     },
     toggleDisabled() {
-      this.info && this.classification && this.classification !== "סודי ביותר"
-        ? (this.disabled = false)
-        : (this.disabled = true);
+      if (this.classification) {
+        let currClassification = this.classifications.filter(
+          (classification) => this.classification == classification.text
+        )[0];
+        this.isClassificationEnable = currClassification?.enable == undefined ? true : currClassification.enable;
+      } else {
+        this.isClassificationEnable = false;
+      }
+      this.info && this.isClassificationEnable ? (this.disabled = false) : (this.disabled = true);
     },
   },
 };
 </script>
 
 <style scoped>
-.top-secret {
+.select-classfier {
+  max-width: 150px;
+}
+.select-classfier-max {
+  max-width: 250px;
+}
+.disable-classification {
   padding-top: 10px;
   color: red;
 }