+++ src/components/popups/external-transfer-popup/steps/Approval.vue
@@ -1,43 +1,105 @@
 <template>
   <div>
-    <div v-if="user.approvalInfo.canApprove">
-      <p class="popup-text align-center">
-        {{ $t("externalTransfer.CanApprove") }}
-      </p>
+    <!-- if the user is admin -->
+    <div v-if="user.approverInfos[networkDest] && user.approverInfos[networkDest].isAdmin">
+      <v-row class="ma-1" no-gutters justify="center">
+        <p class="pa-0 popup-text align-center">
+          {{ $t("externalTransfer.CanApprove") }}
+        </p>
+        <v-btn text small @click="onAboutMeClick">
+          <p>{{ $t("buttons.AboutMe") }}</p>
+        </v-btn>
+      </v-row>
     </div>
+
+    <!-- if the user has no unit and he isnt approver -->
+    <div
+      v-else-if="
+        user.approverInfos[networkDest] &&
+          user.approverInfos[networkDest].noUnit &&
+          !user.approverInfos[networkDest].isApprover
+      "
+    >
+      <v-row class="ma-1" no-gutters justify="center">
+        <p class="pa-0 popup-text align-center">
+          {{ $t("externalTransfer.NoUnit") }}
+        </p>
+
+        <v-btn text small @click="onAboutMeClick">
+          <p>{{ $t("buttons.AboutMe") }}</p>
+        </v-btn>
+      </v-row>
+    </div>
+
+    <!-- if the user is blocked -->
+    <div v-else-if="user.approverInfos[networkDest] && user.approverInfos[networkDest].isBlocked">
+      <v-row class="ma-1" no-gutters justify="center">
+        <p class="pa-0 popup-text align-center">
+          {{ $t("externalTransfer.IsBlocked") }}
+        </p>
+        <v-btn text small @click="onAboutMeClick">
+          <p>{{ $t("buttons.AboutMe") }}</p>
+        </v-btn>
+      </v-row>
+    </div>
+
+    <!-- if the user is approver -->
+    <div v-else-if="user.approverInfos[networkDest] && user.approverInfos[networkDest].isApprover">
+      <v-row class="ma-1" no-gutters justify="center">
+        <p class="pa-0 popup-text align-center">
+          {{ $t("externalTransfer.CanApprove") }}
+        </p>
+        <v-btn text small @click="onAboutMeClick">
+          <p>{{ $t("buttons.AboutMe") }}</p>
+        </v-btn>
+      </v-row>
+    </div>
+
+    <!-- else -->
     <div v-else>
       <div>
-        <div id="approval-header" class="space-between">
-          <p class="popup-text">{{ $t("externalTransfer.ApprovalChoose") }}</p>
-          <v-tooltip top color="#2c3448">
-            <template v-slot:activator="{ on }">
-              <v-icon color="#2c3448" v-on="on">info</v-icon>
-            </template>
+        <div id="approval-header">
+          <v-row class="ma-1 popup-text" no-gutters justify="center" align="center">
+            <p>{{ $t("externalTransfer.ApprovalChoose") }}</p>
+            <v-spacer />
+            <v-btn class="mx-1" rounded color="#2c3448" dark x-small @click="onAboutMeClick">
+              <v-icon small>person</v-icon>
+              <p>{{ $t("buttons.AboutMe") }}</p>
+            </v-btn>
 
-            <div class="align-center">
-              <p>{{ $t("externalTransfer.ApprovalInstructions") }}</p>
-              <div v-if="!user.approvalInfo.requestFaild">
-                <p v-if="user.approvalInfo.unit">
-                  {{ $t("externexternalTransferalShare.ApproverUnit") }}
-                  <span class="bold">{{ user.approvalInfo.unit }}</span>
-                </p>
-                <p v-if="user.approvalInfo.ranks">
-                  , {{ $t("externalTransfer.ApproverRanks") }}
-                  <span class="bold">{{ getRanks() }}</span>
-                </p>
-                <p class="bold">{{ whiteListText }}</p>
+            <v-tooltip top color="#2c3448">
+              <template v-slot:activator="{ on }">
+                <v-icon color="#2c3448" v-on="on">info</v-icon>
+              </template>
+
+              <div class="align-center">
+                <p>{{ $t("externalTransfer.ApprovalInstructions") }}</p>
+                <div v-if="user.approverInfos[networkDest] && !user.approverInfos[networkDest].requestFaild">
+                  <!-- the approver must be in unit msg -->
+                  <p v-if="user.approverInfos[networkDest] && user.approverInfos[networkDest].unit.name">
+                    {{ $t("externalTransfer.ApproverUnit") }}
+                    <span class="bold">{{ user.approverInfos[networkDest].unit.name }}</span>
+                  </p>
+
+                  <!-- the approver must be in unit msg -->
+                  <p v-if="user.approverInfos[networkDest] && user.approverInfos[networkDest].unit.approvers">
+                    {{ $t("externalTransfer.ApproverRanks") }}
+                    <span class="bold">{{ getRanks(user.approverInfos[networkDest].unit.approvers) }}</span>
+                  </p>
+                  <p class="bold">{{ whiteListText }}</p>
+                </div>
               </div>
-            </div>
-          </v-tooltip>
+            </v-tooltip>
+          </v-row>
         </div>
 
         <v-checkbox
-          v-if="user.approvalInfo.requestFaild"
+          v-if="user.approverInfos[networkDest] && user.approverInfos[networkDest].requestFaild"
           class="space-right"
           :disabled="selectedApprovals.length !== 0"
           :label="$t('externalTransfer.IAmApprover')"
           color="#035c64"
-          v-model="canApprove"
+          v-model="iAmApprover"
         ></v-checkbox>
 
         <Autocomplete
@@ -51,30 +113,43 @@
           @type="getUsersByName"
         />
       </div>
+      <div v-if="blockedApprover">
+        <p id="cant-approve">
+          {{
+            $t("externalTransfer.BlockedApprover", {
+              name: blockedApprover.name,
+            })
+          }}
+        </p>
+        <v-btn text small @click="$refs.support.open(blockedApprover)" id="more-info-button">
+          <p>{{ $t("buttons.MoreInfo") }}</p>
+        </v-btn>
+      </div>
+
       <v-chip-group show-arrows>
-        <Chips
-          v-for="user in selectedApprovals"
-          :key="user.id"
-          :user="user"
-          @remove="onRemove"
-        />
+        <Chips v-for="user in selectedApprovals" :key="user.id" :user="user" @remove="onRemove" />
       </v-chip-group>
     </div>
 
     <v-card-actions class="popup-confirm">
-      <SubmitButton
-        @click="onConfirm"
-        :label="$t('buttons.Continue')"
-        :disabled="disabled"
-      />
+      <SubmitButton @click="onConfirm" :label="$t('buttons.Continue')" :disabled="disabled" />
       <TextButton @click="$emit('back')" :label="$t('buttons.Back')" />
+
+      <v-btn text small @click="onAboutMeClick">
+        <v-icon color="black">person</v-icon>
+        <p>{{ $t("buttons.AboutMe") }}</p>
+      </v-btn>
     </v-card-actions>
+
+    <DropboxSupportPopup ref="support" />
   </div>
 </template>
 
 <script>
 import { mapGetters } from "vuex";
 import * as usersApi from "@/api/users";
+
+import DropboxSupportPopup from "@/components/popups/external-transfer-popup/DropboxSupportPopup";
 import Chips from "@/components/shared/BaseChips";
 import Autocomplete from "@/components/inputs/BaseAutocomplete";
 import SubmitButton from "@/components/buttons/BaseSubmitButton";
@@ -83,29 +158,69 @@ import { AdvancedSearchEnum } from "@/utils/advancedSearchEnum";
 
 export default {
   name: "Approval",
-  components: { Chips, SubmitButton, Autocomplete, TextButton },
+  components: {
+    Chips,
+    SubmitButton,
+    Autocomplete,
+    TextButton,
+    DropboxSupportPopup,
+  },
+  props: {
+    networkDest: String,
+    reset: Boolean,
+  },
   data() {
     return {
       users: [],
-      isLoading: false,
       selectedApprovals: [],
+      blockedApprover: undefined,
+      isLoading: false,
       disabled: true,
-      canApprove: false,
+      iAmApprover: false,
     };
   },
   computed: {
     ...mapGetters(["user", "whiteListText"]),
   },
   watch: {
-    selectedApprovals: function (users) {
+    selectedApprovals: function(users) {
       users.length ? (this.disabled = false) : (this.disabled = true);
     },
-    canApprove: function (canApprove) {
-      canApprove ? (this.disabled = false) : (this.disabled = true);
+    reset() {
+      this.users = [];
+      this.selectedApprovals = [];
+      this.blockedApprover = undefined;
+      this.isLoading = false;
+      this.iAmApprover = false;
+
+      this.disabled =
+        this.networkDest == undefined ||
+        !(
+          this.user.approverInfos[this.networkDest].isAdmin ||
+          (this.user.approverInfos[this.networkDest].isApprover && !this.user.approverInfos[this.networkDest].isBlocked)
+        );
+    },
+    networkDest: function() {
+      this.disabled =
+        this.networkDest == undefined ||
+        !(
+          this.user.approverInfos[this.networkDest].isAdmin ||
+          (this.user.approverInfos[this.networkDest].isApprover && !this.user.approverInfos[this.networkDest].isBlocked)
+        );
+    },
+    iAmApprover: function(value) {
+      value ? (this.disabled = false) : (this.disabled = true);
     },
   },
   created() {
-    this.disabled = !this.user.approvalInfo.canApprove;
+    this.disabled =
+      this.networkDest == undefined ||
+      !(
+        this.user.approverInfos[this.networkDest].isAdmin ||
+        (this.user.approverInfos[this.networkDest].isApprover && !this.user.approverInfos[this.networkDest].isBlocked)
+      )
+        ? true
+        : false;
   },
   methods: {
     getUsersByName(name) {
@@ -124,23 +239,48 @@ export default {
         this.selectedApprovals.map((user) => user.id)
       );
     },
-    getRanks() {
-      return this.user.approvalInfo.ranks.toString().split(",").join(", ");
+    getRanks(ranks) {
+      return ranks
+        .toString()
+        .split(",")
+        .join(", ");
     },
-    onSelect(user) {
+    async onSelect(approver) {
+      this.blockedApprover = undefined;
       this.users = [];
-      if (!user) return;
-      else if (this.isUserExists(this.selectedApprovals, user.id))
-        this.remove(user);
-      else this.selectedApprovals.push(user);
+      if (!approver || this.isUserExists(this.selectedApprovals, approver.id)) {
+        this.disabled = true;
+        return;
+      } else {
+        const canApprove = await usersApi.canBeApproved(this.user.id, approver.id, this.networkDest);
+
+        if (canApprove.canApproveToUser) {
+          this.selectedApprovals.push(approver);
+          this.disabled = false;
+        } else if (canApprove.cantApproveReasons) {
+          this.blockedApprover = {
+            name: approver.fullName,
+            reasons: canApprove.cantApproveReasons,
+          };
+          if (this.selectedApprovals.length == 0) {
+            this.disabled = true;
+          }
+        }
+      }
     },
     onRemove(item) {
       this.selectedApprovals = this.selectedApprovals.filter((user) => {
         return user.id !== item.id;
       });
+      if (this.selectedApprovals.length == 0) {
+        this.disabled = true;
+      }
     },
     isUserExists(users, id) {
       return users.some((user) => user.id === id);
+    },
+    onAboutMeClick() {
+      usersApi.openAboutMePage(this.networkDest);
     },
   },
 };
@@ -149,5 +289,13 @@ export default {
 <style scoped>
 #approval-header {
   padding: 3px;
+}
+#cant-approve {
+  color: red;
+  text-align: center;
+}
+#more-info-button {
+  margin: auto;
+  display: block;
 }
 </style>