+++ CHANGELOG.md
@@ -5,17 +5,44 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
-### Added
+## [unreleased]
 - major: ([66](https://github.com/meateam/api-gateway/pull/66)): add advance search by mail/T.
 
+## [v3.3.0] - 2021-03-14
+
+### Added
+- minor([67](https://github.com/meateam/drive-ui/pull/67)): added cts 
+- minor([68](https://github.com/meateam/drive-ui/pull/68)): added button for local office.
+
+## [v3.2.0] - 2021-02-28
+
+### Fixed
+
+- minor([65](https://github.com/meateam/drive-ui/pull/65)): fix all bugs and add approval info.
+
+## [v3.1.1]
+
+### Added
+
+- info about the approvers in the ui.
+
+## Added
+
+- info about the approvers in the ui.
+## [v2.0.2] - 2020-10-28
+
+### Removed
+
+- default timoout for the axios request because some big files uploads that took more then 10 sec faild.
 ## [v2.3.0] - 2020-12-24
+
 ### Added
 
 - add popup for convert docs.
 
 - add pagination to shared files and external transferd.
+- 
 ## [v2.2.1]
 
 ### Fixed
@@ -59,6 +86,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - default timoout for the axios request because some big files uploads that took more then 10 sec faild.
+
+[unreleased]: https://github.com/meateam/api-gateway/compare/master...develop
+[v3.3.0]: https://github.com/meateam/api-gateway/compare/v3.2.0...v3.3.0
+[v3.2.0]: https://github.com/meateam/api-gateway/compare/v3.1.1...v3.2.0
+[v3.1.1]: https://github.com/meateam/api-gateway/compare/v2.3.0...v3.1.1
+[v2.3.0]: https://github.com/meateam/api-gateway/compare/v2.2.1...v2.3.0
+[v2.2.1]: https://github.com/meateam/api-gateway/compare/v2.2.0...v2.2.1
+[v2.2.0]: https://github.com/meateam/api-gateway/compare/v2.1.1...v2.2.0
+[v2.1.1]: https://github.com/meateam/api-gateway/compare/v2.1.0...v2.1.1
+[v2.1.0]: https://github.com/meateam/api-gateway/compare/v2.0.2...v2.1.0
+[v2.0.2]: https://github.com/meateam/api-gateway/compare/develop...v2.0.2